+++ lib/active_record/connection_adapters/percona_adapter.rb
@@ -130,6 +130,20 @@ module ActiveRecord
         true
       end
 
+      # Executes the passed statement through pt-online-schema-change if it's
+      # an alter statement, or through the mysql adapter otherwise
+      #
+      # @param sql [String]
+      # @param name [String]
+      def percona_execute(sql, name)
+        if alter_statement?(sql)
+          command = cli_generator.parse_statement(sql)
+          runner.execute(command)
+        else
+          mysql_adapter.execute(sql, name)
+        end
+      end
+
       # This abstract method leaves up to the connection adapter freeing the
       # result, if it needs to. Check out: https://github.com/rails/rails/blob/330c6af05c8b188eb072afa56c07d5fe15767c3c/activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb#L247
       #
@@ -142,6 +156,14 @@ module ActiveRecord
       private
 
       attr_reader :mysql_adapter, :logger, :runner, :cli_generator
+
+      # Checks whether the sql statement is an ALTER TABLE
+      #
+      # @param sql [String]
+      # @return [Boolean]
+      def alter_statement?(sql)
+        sql =~ /alter table/i
+      end
     end
   end
 end