+++ spec/active_record/connection_adapters/percona_adapter_spec.rb
@@ -282,4 +282,61 @@ describe ActiveRecord::ConnectionAdapters::PerconaMigratorAdapter do
       adapter.create_table(table_name)
     end
   end
+
+  describe '#percona_execute' do
+    let(:name) { nil }
+
+    context 'when an alter statement is provided' do
+      let(:table_name) { :comments }
+      let(:statement) do
+        'ALTER TABLE `comments` CHANGE `some_id` `some_id` INT(11) DEFAULT NULL'
+      end
+
+      before do
+        allow(cli_generator).to(
+          receive(:parse_statement).with(statement)
+        ).and_return('percona command')
+      end
+
+      it 'passes the built SQL to the CliGenerator' do
+        expect(cli_generator).to receive(:parse_statement).with(statement)
+        adapter.percona_execute(statement, name)
+      end
+
+      it 'runs the command' do
+        expect(runner).to receive(:execute).with('percona command')
+        adapter.percona_execute(statement, name)
+      end
+
+      context 'and providing a name' do
+        let(:name) { 'name' }
+
+        it 'passes the built SQL to the CliGenerator' do
+          expect(cli_generator).to receive(:parse_statement).with(statement)
+          adapter.percona_execute(statement, name)
+        end
+
+        it 'runs the command' do
+          expect(runner).to receive(:execute).with('percona command')
+          adapter.percona_execute(statement, name)
+        end
+      end
+    end
+
+    context 'when a non-alter statement is provided' do
+      let(:statement) { 'UPDATE comments SET some_id = NULL' }
+
+      it 'delegates to the mysql adapter' do
+        expect(mysql_adapter).to receive(:execute).with(statement, nil)
+        adapter.percona_execute(statement, nil)
+      end
+
+      context 'and providing a name' do
+        it 'delegates to the mysql adapter' do
+          expect(mysql_adapter).to receive(:execute).with(statement, 'name')
+          adapter.percona_execute(statement, 'name')
+        end
+      end
+    end
+  end
 end